Add client-side search
Using the Lunr JS library, we can search through all of the blogs and using client-side JS
selectively hide/show only the blogs that match the query. The current implementation only
searches by the blog title & author.

When there are no results, the page renders like normal but with a red "No results found"
line added under the new search bar at the top.

When there are results, all the blogs are hidden before only the relevant ones are
re-shown. This is to avoid having to loop over every single blog (a quite large, and still growing
number) to decide show/hide. Other elements (like the table of contents, and empty category
headers) are also hidden as appropriate.

[1] https://lunrjs.com